+++ Runtime/Serie/Line/LineHandler.PolarCoord.cs
@@ -47,21 +47,19 @@ namespace XCharts.Runtime
             }
             m_LastCheckContextFlag = needCheck;
             var themeSymbolSize = chart.theme.serie.lineSymbolSize;
-            var themeSymbolSelectedSize = chart.theme.serie.lineSymbolSelectedSize;
             lineWidth = serie.lineStyle.GetWidth(chart.theme.serie.lineWidth);
 
             var needInteract = false;
             if (m_LegendEnter)
             {
+                serie.context.pointerEnter = true;
                 serie.interact.SetValue(ref needInteract, lineWidth, true, chart.theme.serie.selectedRate);
                 for (int i = 0; i < serie.dataCount; i++)
                 {
                     var serieData = serie.data[i];
-                    var symbol = SerieHelper.GetSerieSymbol(serie, serieData);
-                    var symbolSelectedSize = symbol.GetSelectedSize(serieData.data, themeSymbolSelectedSize);
-
+                    var size = SerieHelper.GetSysmbolSize(serie, serieData, chart.theme, themeSymbolSize, SerieState.Emphasis);
                     serieData.context.highlight = true;
-                    serieData.interact.SetValue(ref needInteract, symbolSelectedSize);
+                    serieData.interact.SetValue(ref needInteract, size);
                 }
             }
             else
@@ -205,27 +203,23 @@ namespace XCharts.Runtime
                     continue;
 
                 var count = serie.dataCount;
-                Color32 symbolColor, symbolToColor,symbolEmptyColor;
+                float symbolBorder = 0f;
+                float[] cornerRadius = null;
+                Color32 symbolColor, symbolToColor, symbolEmptyColor, borderColor;
                 for (int i = 0; i < count; i++)
                 {
                     var serieData = serie.GetSerieData(i);
-                    var symbol = SerieHelper.GetSerieSymbol(serie, serieData);
+                    var state = SerieHelper.GetSerieState(serie, serieData, true);
+                    var symbol = SerieHelper.GetSerieSymbol(serie, serieData, state);
                     if (ChartHelper.IsIngore(serieData.context.position))
                         continue;
 
-                    bool highlight = serieData.context.highlight || serie.highlight;
-                    if ((!symbol.show || !symbol.ShowSymbol(i, count) || serie.IsPerformanceMode()) &&
-                        !serieData.context.highlight)
+                    if (!symbol.show || !symbol.ShowSymbol(i, count))
                         continue;
 
-                    var symbolSize = highlight ?
-                        symbol.GetSelectedSize(serieData.data, chart.theme.serie.lineSymbolSize) :
-                        symbol.GetSize(serieData.data, chart.theme.serie.lineSymbolSize);
-                    var state = SerieHelper.GetSerieState(serie, serieData);
-                    SerieHelper.GetItemColor(out symbolColor, out symbolToColor,out symbolEmptyColor, serie, serieData, chart.theme, n);
-                    var symbolBorder = SerieHelper.GetSymbolBorder(serie, serieData, chart.theme, state);
-                    var borderColor = SerieHelper.GetSymbolBorderColor(serie, serieData, chart.theme, state);
-                    var cornerRadius = SerieHelper.GetSymbolCornerRadius(serie, serieData, state);
+                    var symbolSize = SerieHelper.GetSysmbolSize(serie, serieData, chart.theme, chart.theme.serie.lineSymbolSize, state);
+                    SerieHelper.GetItemColor(out symbolColor, out symbolToColor, out symbolEmptyColor, serie, serieData, chart.theme, n);
+                    SerieHelper.GetSymbolInfo(out borderColor, out symbolBorder, out cornerRadius, serie, null, chart.theme, state);
 
                     symbolSize = serie.animation.GetSysmbolSize(symbolSize);
                     chart.DrawSymbol(vh, symbol.type, symbolSize, symbolBorder, serieData.context.position,