+++ Runtime/Serie/Scatter/BaseScatterHandler.cs
@@ -82,22 +82,21 @@ namespace XCharts.Runtime
             for (int i = serie.dataCount - 1; i >= 0; i--)
             {
                 var serieData = serie.data[i];
-                var symbol = SerieHelper.GetSerieSymbol(serie, serieData);
-                var symbolSize = symbol.GetSize(serieData.data, themeSymbolSize);
-                var symbolSelectedSize = symbol.GetSelectedSize(serieData.data, themeSymbolSelectedSize);
+                var symbolSize = SerieHelper.GetSysmbolSize(serie, serieData, chart.theme, themeSymbolSize);
                 if (m_LegendEnter ||
                     (!needHideAll && Vector3.Distance(serieData.context.position, chart.pointerPos) <= symbolSize))
                 {
                     serie.context.pointerItemDataIndex = i;
                     serie.context.pointerEnter = true;
                     serieData.context.highlight = true;
-                    serieData.interact.SetValue(ref needInteract, symbolSelectedSize);
                 }
                 else
                 {
                     serieData.context.highlight = false;
-                    serieData.interact.SetValue(ref needInteract, symbolSize);
                 }
+                var state = SerieHelper.GetSerieState(serie, serieData, true);
+                symbolSize = SerieHelper.GetSysmbolSize(serie, serieData, chart.theme, themeSymbolSize, state);
+                serieData.interact.SetValue(ref needInteract, symbolSize);
             }
             if (needInteract)
             {
@@ -144,19 +143,19 @@ namespace XCharts.Runtime
             serie.containerIndex = m_Grid.index;
             serie.containterInstanceId = m_Grid.instanceId;
 
+            float symbolBorder = 0f;
+            float[] cornerRadius = null;
+            Color32 color, toColor, emptyColor, borderColor;
             foreach (var serieData in dataList)
             {
                 var symbol = SerieHelper.GetSerieSymbol(serie, serieData);
                 if (!symbol.ShowSymbol(serieData.index, maxCount))
                     continue;
 
-                var highlight = serie.highlight || serieData.context.highlight;
-                var state = SerieHelper.GetSerieState(serie, serieData);
-                Color32 color, toColor, emptyColor;
+                var state = SerieHelper.GetSerieState(serie, serieData, true);
+
                 SerieHelper.GetItemColor(out color, out toColor, out emptyColor, serie, serieData, chart.theme, colorIndex, state);
-                var symbolBorder = SerieHelper.GetSymbolBorder(serie, serieData, theme, state);
-                var borderColor = SerieHelper.GetSymbolBorderColor(serie, serieData, theme, state);
-                var cornerRadius = SerieHelper.GetSymbolCornerRadius(serie, serieData, state);
+                SerieHelper.GetSymbolInfo(out borderColor, out symbolBorder, out cornerRadius, serie, serieData, chart.theme, state);
                 double xValue = serieData.GetCurrData(0, dataChangeDuration, xAxis.inverse);
                 double yValue = serieData.GetCurrData(1, dataChangeDuration, yAxis.inverse);
 
@@ -175,14 +174,10 @@ namespace XCharts.Runtime
                 serie.context.dataPoints.Add(pos);
                 serieData.context.position = pos;
                 var datas = serieData.data;
-                var symbolSize = serie.highlight || serieData.context.highlight ?
-                    theme.serie.scatterSymbolSelectedSize :
-                    theme.serie.scatterSymbolSize;
+                var symbolSize = 0f;
                 if (!serieData.interact.TryGetValue(ref symbolSize, ref interacting))
                 {
-                    symbolSize = highlight ?
-                        symbol.GetSelectedSize(serieData.data, symbolSize) :
-                        symbol.GetSize(serieData.data, symbolSize);
+                    symbolSize = SerieHelper.GetSysmbolSize(serie, serieData, chart.theme, chart.theme.serie.scatterSymbolSize, state);
                     serieData.interact.SetValue(ref interacting, symbolSize);
                 }
 
@@ -249,24 +244,25 @@ namespace XCharts.Runtime
             serie.containerIndex = axis.index;
             serie.containterInstanceId = axis.instanceId;
 
+            float symbolBorder = 0f;
+            float[] cornerRadius = null;
+            Color32 color, toColor, emptyColor, borderColor;
             foreach (var serieData in dataList)
             {
                 var symbol = SerieHelper.GetSerieSymbol(serie, serieData);
                 if (!symbol.ShowSymbol(serieData.index, maxCount))
                     continue;
 
-                var state = SerieHelper.GetSerieState(serie, serieData);
-                Color32 color, toColor, emptyColor;
+                var state = SerieHelper.GetSerieState(serie, serieData, true);
                 SerieHelper.GetItemColor(out color, out toColor, out emptyColor, serie, serieData, chart.theme, colorIndex, state);
-                var symbolBorder = SerieHelper.GetSymbolBorder(serie, serieData, theme, state);
-                var borderColor = SerieHelper.GetSymbolBorderColor(serie, serieData, theme, state);
-                var cornerRadius = SerieHelper.GetSymbolCornerRadius(serie, serieData, state);
-                var xValue = serieData.GetCurrData(0, dataChangeDuration, axis.inverse);
+                SerieHelper.GetSymbolInfo(out borderColor, out symbolBorder, out cornerRadius, serie, serieData, chart.theme, state);
 
                 if (serieData.IsDataChanged())
                     dataChanging = true;
 
                 var pos = Vector3.zero;
+                var xValue = serieData.GetCurrData(0, dataChangeDuration, axis.inverse);
+
                 if (axis.orient == Orient.Horizonal)
                 {
                     var xDataHig = GetDataHig(axis, xValue, axis.context.width);
@@ -283,11 +279,7 @@ namespace XCharts.Runtime
                 serieData.context.position = pos;
 
                 var datas = serieData.data;
-                var symbolSize = 0f;
-                if (serie.highlight || serieData.context.highlight)
-                    symbolSize = symbol.GetSelectedSize(datas, theme.serie.scatterSymbolSelectedSize);
-                else
-                    symbolSize = symbol.GetSize(datas, theme.serie.scatterSymbolSize);
+                var symbolSize = SerieHelper.GetSysmbolSize(serie, serieData, chart.theme, chart.theme.serie.scatterSymbolSize, state);
                 symbolSize *= rate;
 
                 if (isEffectScatter)